Add files via upload
This Presentation of mine focuses on how I had implemented a new solution to our department which included implementation of software and technique to our department.
It also became so efficient in a way that it helped our department saving more than 70% of time and reduced efforts. I could also help save 80% of cost on expenses of our department and I could predict all the graduating rates of our students in a timely manner. 
This project was done in Fall 2022 at our university and it impacted to implement in other departments as well.